test(mitm-addon): split firewall rewrite tests
